compiling Arduino code

+++ arduino/making_sense/sensors.h
@@ -29,6 +29,6 @@ inline float c2f(float c);
 inline float d2f(int16_t d);
 float readTempSensor  (uint8_t pin);
 
-uint16_t readPressSensor (uint8_t pin);
+int readPressSensor (uint8_t pin);
 
 #endif /* _SENSORS_H_ */
